refactor: fix navbar usage
